feat: add security to protect product + profile management

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, reverse, redirect, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from .models import Product, Category
 from .forms import ProductForm
@@ -52,7 +53,14 @@ class ProductDetail(generic.DetailView):
     context_object_name = "product"
 
 
+@login_required
 def add_product(request):
+    if not request.user.is_superuser:
+        messages.error(
+            request, "You need proper authorisation to visit this page."
+        )
+        return redirect(reverse("home"))
+
     if request.method == "POST":
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -75,7 +83,14 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def update_product(request, product_id):
+    if not request.user.is_superuser:
+        messages.error(
+            request, "You need proper authorisation to visit this page."
+        )
+        return redirect(reverse("home"))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == "POST":
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -100,7 +115,14 @@ def update_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
+    if not request.user.is_superuser:
+        messages.error(
+            request, "You need proper authorisation to visit this page."
+        )
+        return redirect(reverse("home"))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, "Successfully deleted product!")
